Refactor to support PHPUnit's new event system
This change refactors some of the test code to support PHPUnit's most
recent iteration of its event system. The previous iteration worked for
earlier versions, but there was a newer version release in PHPUnit
10.0.7.
