+++ phpunit.xml.dist
@@ -4,11 +4,10 @@
     xsi:noNamespaceSchemaLocation="vendor/phpunit/phpunit/phpunit.xsd"
     bootstrap="vendor/autoload.php"
     colors="true">
-    <coverage processUncoveredFiles="true">
-        <include>
-            <directory suffix=".php">./src</directory>
-        </include>
-    </coverage>
+
+    <extensions>
+        <bootstrap class="LaminasIntegrationTest\Db\Extension\ListenerExtension" />
+    </extensions>
 
     <testsuites>
         <testsuite name="unit test">
@@ -19,9 +18,6 @@
         </testsuite>
     </testsuites>
 
-    <listeners>
-        <listener class="LaminasIntegrationTest\Db\IntegrationTestListener" file="./test/integration/IntegrationTestListener.php"/>
-    </listeners>
 
     <php>
         <!-- Note: the following is a FULL list of ALL POSSIBLE constants
