+++ test/integration/Extension/ListenerExtension.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace LaminasIntegrationTest\Db\Extension;
+
+use PHPUnit\Runner\Extension\Extension;
+use PHPUnit\Runner\Extension\Facade;
+use PHPUnit\Runner\Extension\ParameterCollection;
+use PHPUnit\TextUI\Configuration\Configuration;
+
+final class ListenerExtension implements Extension
+{
+    public function bootstrap(
+        Configuration $configuration,
+        Facade $facade,
+        ParameterCollection $parameters
+    ): void {
+        $facade->registerSubscribers(
+            new IntegrationTestStartedListener(),
+            new IntegrationTestStoppedListener(),
+        );
+    }
+}
